+++ app/(ui)/score/page/[page]/page.tsx
@@ -1,4 +1,4 @@
-import Scores from './Scores'
+import Scores from './../../Scores'
 import {verifySession} from '@/app/lib/session'
 import {redirect} from "next/navigation";
 import {Routes} from "@routes";
@@ -6,12 +6,15 @@ import {fetchScores} from "@/app/lib/data";
 
 export const revalidate = 0;
 
+interface PageProps {
+  params: { page: string };
+}
 
-export default async function Page() {
+export default async function Page({params}: PageProps) {
   const session = await verifySession();
 
-  const { scores, total } = await fetchScores(1, 18);
-
+  const page = parseInt(params.page, 10) || 1;
+  const { scores, total } = await fetchScores(page, 18);
 
   if (session) {
     const currentUserId = session.userId;
